reorganize file structure again... Add basic example of a python module connecting to the network
