+++ manage
@@ -3,9 +3,8 @@
 set -e
 
 SCRIPT_HOME="$( cd "$( dirname "$0" )" && pwd )"
-DOCKER_DIRECTORY="docker"
 
-cd "$SCRIPT_HOME/$DOCKER_DIRECTORY"
+cd "$SCRIPT_HOME"
 
 case "$1" in
   start_nodes)
@@ -13,6 +12,9 @@ case "$1" in
     ;;
   start_client)
       docker-compose run client
+    ;;
+  start_von_connector)
+      docker-compose run von_connector
     ;;
   stop)
       docker-compose stop
