+++ scripts/start_node.sh
