+++ scripts/start_von_connector.sh
@@ -1,5 +1,9 @@
 #! /bin/bash
 
+set -e
+
+SCRIPT_HOME="$( cd "$( dirname "$0" )" && pwd )"
+
 PATH="$PATH:/home/indy/.local/bin"
 
 # Get ip addresses from docker container names
@@ -7,6 +11,11 @@ IP1="$(getent hosts node1 | awk '{ print $1 }')"
 IP2="$(getent hosts node2 | awk '{ print $1 }')"
 IP3="$(getent hosts node3 | awk '{ print $1 }')"
 IP4="$(getent hosts node4 | awk '{ print $1 }')"
+
+if [ -z "$IP1" ] || [ -z "$IP2" ] || [ -z "$IP3" ] || [ -z "$IP4" ]; then
+    echo "Cannot discover node ips. Are the nodes running?"
+    exit 1
+fi
 
 echo generate_indy_pool_transactions \
     --nodes 4 \
@@ -26,4 +35,4 @@ pipenv --three
 # `pipenv install --pre python3-indy` should install latest dev
 # release but it isn't working... Using latest dev build (244)
 pipenv install python3-indy==1.0.1-dev-244
-pipenv shell
+pipenv run python ./von-connector/main.py
